backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-generativeai==0.4.0
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic==2.6.3
python-multipart==0.0.9
